requirements.txt
gTTS-2.4.0 
google-ai-generativelanguage-0.4.0 
google-generativeai-0.3.2 
pyperclip-1.8.2 
python-docx-1.0.1 
python-dotenv-1.0.0 
requests-2.31.0 
streamlit-1.31.0